refactor: separate ytmusic integration tests
